+++ engine/src/Rendering/Renderer.cpp
@@ -1651,7 +1651,7 @@ void Renderer::ReallocateRenderObjects(unsigned int required)
 	if (required <= data.allocated)
 		return;
 
-	required = Math::UpperPowerOfTwo(required);
+	required = static_cast<unsigned int>(Math::UpperPowerOfTwo(required));
 
 	// Reserve same amount in entity map
 	entityMap.Reserve(required);
